Report rig backlog and failures honestly.
Addresses the automated review of #478 (6 fix, 2 document, 4 consider).

Taken:

- build-ryll.sh is now executable, as the rerun instructions assume.
- summarise.py's backend column was skmem r (kernel memory, payload
  plus per-skb overhead), labelled as unread bytes. It now reports
  Recv-Q as the backlog and skmem r separately as memory. Re-derived
  from the preserved phase 1 results: the stock rows barely move
  (80 ms / 10 Mbit heavy 1954 -> 1835 KiB) but the capped rows do
  (386 -> 255-292, 93 -> 28), so the proxy-backpressure table and its
  prose are corrected. That prose was already wrong at 80 ms / 50 Mbit,
  where tuned/small leave 0.9-1.2 MiB rather than "0.3-0.7".
- summarise.py pools only good repeats and reports good / qemu-died /
  failed per cell, with a stderr warning; unreadable repeats count as
  failed instead of being silently skipped.
- run_case records setup failures (qemu port, proxy start exit status,
  ryll socket or ryll exit) in a setup-failed marker and stops what it
  started; before, the early return leaked qemu and the other two fell
  through to thin data.
- keydraw-latency.py sends an unmeasured probe press after warmup and
  requires a key-box draw, so a missing rect patch or a dead guest is
  a named failure even on an idle guest. The old guard exited 0 with
  only timeouts in both cases.
- Both tuning settings are bounded at 2**32 - 1 in config.py, and both
  proxy flags are u32, so a config value the daemon accepts is never
  one clap rejects.
- The SharedState construction is a function, with a test that the
  flags reach it and that both flags share the config's bound.
- The SO_RCVBUF test allows for net.core.rmem_max.
- summarise.py prints the /metrics relay rate next to the ss rate,
  which makes the page's agreement claim checkable; the unused
  notsent/unacked parsing is gone.
- docs/testing.md gains a shaped-link rig subsection, and
  TCP_NOTSENT_LOWAT is documented as applying to TLS client sockets
  only.

Declined: item 8 (event-driven activity start in the guest). The
recorded data has no qemu-died repeats, and a crash from the ordering
inverting is now visible per cell rather than averaged in.

Ten mutations, all caught: qemu-died, setup-failed and unreadable
repeats; Recv-Q versus skmem; the config bound; three probe outcomes
against a fake control socket; SharedState wiring (Docker).

Prompt: Michael asked for the automated review on kerbside#478 to be
addressed through the whole respond-to-review cycle.

+++ rust/kerbside-proxy/src/session.rs
@@ -132,7 +132,7 @@ pub struct SharedState {
     pub sessions: Arc<SessionRegistry>,
     /// `SO_RCVBUF` cap applied to each backend-leg socket once connected
     /// (`--backend-rcvbuf-bytes`); 0 leaves the kernel's autotuning alone.
-    pub backend_rcvbuf_bytes: usize,
+    pub backend_rcvbuf_bytes: u32,
 }
 
 /// Overall time budget for the client-facing handshake reads/writes (link
